Refactore la duplication de restitutions des questions et réponses dans une classe à part
Dans Livraison et Bienvenue
