+++ app/assets/stylesheets/restitution_globale/_base.scss
@@ -117,10 +117,6 @@
     @include applique-theme($couleur-questions);
   }
 
-  .reponse-communication-ecrite {
-    margin-top: .5rem;
-  }
-
   .competences-identifiees {
     list-style: none;
     display: flex;
