+++ app/models/restitution/livraison.rb
@@ -4,10 +4,6 @@ require_relative '../../decorators/evenement_livraison'
 
 module Restitution
   class Livraison < AvecEntrainement
-    EVENEMENT = {
-      REPONSE: 'reponse'
-    }.freeze
-
     METRIQUES = {
       'nombre_bonnes_reponses_numeratie' => {
         'type' => :nombre,
@@ -54,39 +50,12 @@ module Restitution
     end
 
     def questions_et_reponses
-      questions_qcm_repondues
-        .map do |question|
-        {
-          question: question,
-          reponse: trouve_reponse(question)
-        }
-      end
-    end
-
-    def reponses
-      evenements_situation.find_all { |e| e.nom == EVENEMENT[:REPONSE] }
+      qr = QuestionsReponses.new(evenements_situation, situation.questionnaire)
+      qr.questions_et_reponses
     end
 
     def efficience
       nil
-    end
-
-    private
-
-    def questions
-      situation.questionnaire.questions
-    end
-
-    def questions_qcm_repondues
-      questions_ids = reponses.collect { |r| r.donnees['question'] }
-      questions.where(id: questions_ids, type: 'QuestionQcm')
-    end
-
-    def trouve_reponse(question)
-      reponse_id = reponses.find do |evenement|
-        evenement.donnees['question'] == question.id
-      end.donnees['reponse']
-      question.choix.find reponse_id
     end
   end
 end
