+++ app/views/admin/evaluations/_bienvenue.arb
@@ -2,14 +2,17 @@
 
 div class: 'row questions' do
   restitution.questions_et_reponses.each do |question_et_reponse|
+    reponse = question_et_reponse[:reponse]
+    question = question_et_reponse[:question]
+
     div class: 'col-4 px-5 mb-4' do
-      div strong question_et_reponse[:question].libelle
-      reponse = if question_et_reponse[:question].is_a?(QuestionQcm)
-                  question_et_reponse[:question].choix.find(question_et_reponse[:reponse]).intitule
+      div strong question.libelle
+      reponse = if question.is_a?(QuestionQcm)
+                  reponse.intitule
                 else
-                  question_et_reponse[:reponse]
+                  reponse
                 end
-      div simple_format(reponse), class: 'reponse-communication-ecrite'
+      div simple_format(reponse), class: 'my-2'
     end
   end
 end
