+++ app/views/admin/evaluations/_livraison.arb
@@ -2,18 +2,22 @@
 
 div class: 'row questions' do
   restitution.questions_et_reponses.each do |question_et_reponse|
-    choix = question_et_reponse[:reponse]
+    reponse = question_et_reponse[:reponse]
     question = question_et_reponse[:question]
 
     div class: 'col-4 px-5 mb-4' do
       div strong question.libelle
-      div class: 'btn-group my-2' do
-        %w[bon mauvais].each do |nom|
-          active = nom == choix.type_choix ? 'active' : ''
-          button class: "btn #{active}" do
-            t(".#{nom}")
+      if question.is_a?(QuestionQcm)
+        div class: 'btn-group my-2' do
+          %w[bon mauvais].each do |nom|
+            active = nom == reponse.type_choix ? 'active' : ''
+            button class: "btn #{active}" do
+              t(".#{nom}")
+            end
           end
         end
+      else
+        div simple_format(reponse), class: 'my-2'
       end
     end
   end
